refactor(functions): extract validate move
